fix: restore common CallManager interface
