+++ logic/src/commonMain/kotlin/com/wire/kalium/logic/feature/call/usecase/AnswerCallUseCase.kt
@@ -1,7 +1,5 @@
 package com.wire.kalium.logic.feature.call.usecase
 
-import com.wire.kalium.logic.data.call.CallType
-import com.wire.kalium.logic.data.call.ConversationType
 import com.wire.kalium.logic.data.id.ConversationId
 import com.wire.kalium.logic.feature.call.CallManager
